backend/requirements.txt
fastapi[all]
sqlalchemy
pydantic
bcrypt
pyjwt
opencv-python
requests
numpy
pillow